third_party/WebKit/LayoutTests/fast/forms/select-popup/popup-menu-resize-after-open-expected.txt

PASS popupWindow.innerWidth is 100
PASS popupWindow.innerWidth is 150
PASS successfullyParsed is true

TEST COMPLETE
